♻️ #119 - useClickOutside 훅 생성 및 적용
- 더보기 버튼 이외의 영역 클릭시 모달 닫기 기능
